+++ CHANGELOG.md
@@ -13,5 +13,12 @@ All notable changes to the "ara-yaml-assist" extension will be documented in thi
 - Fix the compatibility issue by supporting the VS Code from 1.64.1
 - Small display bug fixes
 
-## 0.0.4 - 2023-01-25
+## 0.0.4 - 2023-01-24
 - Rearrange the changelog and readme files
+
+## 0.0.5 - 2023-01-24
+- Fix some data type prompts
+
+## 0.0.6 - 2023-01-24
+- Add the support to get the current vs editor settings on tabs/spaces (not fixed on 2 spaces)
+- Add hover over framework and a few hover description test